// ang-newsapp/src/app/models/news.ts
export class News {
    SourceName?: string;
    author?: string;
    title?: string;
    description?: string;
    url?: string;
    urltoImage?: string;
    publishedAt?: Date;
    isFavourite?: boolean;
    isRecommend?:boolean;
    id?: number;
    userName?:string;
    
    constructor() {
    }
}